+++ src/softmax.cpp
@@ -13,13 +13,13 @@ namespace Layers {
 Tensor Softmax::apply(const Tensor& input) const {
     Tensor res = input;
     for (int i = 0; i < res.shape[0]; i++) {
-        double max = res.data[0];
+        float max = res.data[0];
         for (int j = 1; j < res.shape[1]; j++) {
             if (res.data[i * res.shape[1] + j] > max) {
                 max = res.data[i * res.shape[1] + j];
             }
         }
-        double sum = 0;
+        float sum = 0;
         for (int j = 0; j < res.shape[1]; j++) {
             res.data[i * res.shape[1] + j] = exp(res.data[i * res.shape[1] + j] - max);
             sum += res.data[i * res.shape[1] + j];
@@ -35,19 +35,19 @@ Tensor Softmax::backward(const Tensor& input_vals, const Tensor& output_grad) {
     Tensor res{input_vals.shape, 0.0, input_vals.device};
     Tensor out = input_vals;
     for (int i = 0; i < res.shape[0]; i++) {
-        double max = out.data[i * res.shape[1]];
+        float max = out.data[i * res.shape[1]];
         for (int j = 1; j < res.shape[1]; j++) {
             if (out.data[i * res.shape[1] + j] > max) {
                 max = out.data[i * res.shape[1] + j];
             }
         }
-        double sum = 0;
+        float sum = 0;
         for (int j = 0; j < res.shape[1]; j++) {
             out.data[i * out.shape[1] + j] = exp(out.data[i * out.shape[1] + j] - max);
             sum += out.data[i * out.shape[1] + j];
         }
 
-        double denom = sum * sum;
+        float denom = sum * sum;
         for (int j = 0; j < res.shape[1]; j++) {
             for (int k = 0; k < res.shape[1]; k++) {
                 if (j == k) {
